requirements.txt
annotated-types==0.5.0
anyio==3.7.1
blinker==1.6.2
click==8.1.7
colorama==0.4.6
Django==2.1
fastapi==0.103.1
Flask==2.3.3  #Micro web framework for Python
Flask-MySQL==1.5.2
Flask-MySQLdb==2.0.0 # Flask-MySQLdb: Flask extension for MySQL integration
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
h11==0.14.0
httptools==0.6.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
mysql-connector-python==9.0.0 #A MySQL driver written in Python, used to connect to MySQL databases
mysqlclient==2.2.4
numpy==1.25.2
pydantic==2.3.0
pydantic_core==2.6.3
pygame==2.5.1
PyMySQL==1.1.1 # PyMySQL: A pure-Python MySQL client library
pypng==0.20220715.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
scikit-learn==1.3.0
scipy==1.11.2
sniffio==1.3.0
SQLAlchemy==2.0.27
starlette==0.27.0
threadpoolctl==3.2.0
typing_extensions==4.7.1
uvicorn==0.23.2
watchfiles==0.20.0
websockets==11.0.3
Werkzeug==2.3.7  # Werkzeug: A comprehensive WSGI web application library (used for security and more)
WTForms==3.1.2

#to install packages
#pip install -r requirements.txt